Clean up of form components
Part of #10713

This change addresses the cleanup of the form components by removing the
Id and Class parameters as explicit parameters. These components will
now propagate the id and class attributes to the created HTML element
without the need for explicit parameters.

Note that we preserve the combining behaviour of class with FieldClass.